Update Integration test lifecycle to be more well defined.
New Lifecycle is as follows.
* `sbt clean it:test`
* sbt will compile and build the one-jar of cosmos
* pre it:test, an in process zk cluster will be started, a temporary directory will be created and cosmos will be started pointing to these
* Integration tests will run, using a single client CosmosIntegrationTestClient (which provides a client to cosmos as well as admin router)
* post it:test, the started zk, cosmos and temporary directory will all be cleaned up.

A few notes:
* std{out,err} from running the cosmos server will be shown in the sbt logs prefixed with `<<cosmos-server>>` on each message
* To run integration tests in your ide, a new system property `com.mesosphere.cosmos.test.CosmosIntegrationTestClient.CosmosClient.uri` will need to be set to point to cosmos
* System properties are still used to initialize the connection to admin router
