+++ src/it/resources/logback-test.xml
@@ -7,19 +7,29 @@
     <encoder>
       <pattern>%d{HH:mm:ss.SSS} %-5level [%thread] %logger{36} - %msg%n</pattern>
     </encoder>
-    <filter class="ch.qos.logback.classic.filter.LevelFilter">
-      <level>TRACE</level>
-      <onMatch>DENY</onMatch>
-      <onMismatch>ACCEPT</onMismatch>
-    </filter>
   </appender>
 
-  <logger name="org.apache.zookeeper" level="error"/>
-  <logger name="org.apache.curator" level="warn"/>
-  <logger name="org.apache.zookeeper.server.quorum.QuorumCnxManager" level="off"/>
-  <logger name="org.apache.zookeeper.ClientCnxn" level="error"/>
+  <appender name="NOEXCEPTION" class="ch.qos.logback.core.ConsoleAppender">
+    <encoder>
+      <pattern>%d{HH:mm:ss.SSS} %-5level [%thread] %logger{36} - %msg%nopex%n</pattern>
+    </encoder>
+  </appender>
 
-  <logger name="access-log" level="trace"/>
+  <logger name="org.apache.zookeeper" level="error" additivity="false">
+    <appender-ref ref="NOEXCEPTION"/>
+  </logger>
+  <logger name="org.apache.curator" level="warn" additivity="false">
+    <appender-ref ref="NOEXCEPTION"/>
+  </logger>
+  <logger name="org.apache.zookeeper.server.quorum.QuorumCnxManager" level="off" additivity="false">
+    <appender-ref ref="NOEXCEPTION"/>
+  </logger>
+  <logger name="org.apache.zookeeper.ClientCnxn" level="error" additivity="false">
+    <appender-ref ref="NOEXCEPTION"/>
+  </logger>
+
+  <!-- Set this above debug to turn off request tracing -->
+  <logger name="com.mesosphere.cosmos.test.CosmosIntegrationTestClient.CosmosClient" level="debug"/>
 
   <root level="debug">
     <appender-ref ref="STDOUT"/>
