+++ src/it/scala/com/mesosphere/cosmos/LocalUniverseSpec.scala
@@ -1,108 +1,115 @@
 package com.mesosphere.cosmos
 
-import cats.data.Xor
-import com.twitter.finagle.http.{FileElement, Method, Request, Status}
-import com.twitter.io.Buf
-import io.circe.parse.parse
-import io.circe.syntax._
-
 import java.io.ByteArrayOutputStream
 import java.nio.file.{Files, Paths}
 import java.util.zip.{ZipEntry, ZipOutputStream}
 
-final class LocalUniverseSpec extends IntegrationSpec {
+import cats.data.Xor
+import com.mesosphere.cosmos.test.CosmosIntegrationTestClient.CosmosClient
+import com.twitter.finagle.http.{FileElement, Method, Request, Status}
+import com.twitter.io.Buf
+import io.circe.parse.parse
+import io.circe.syntax._
+import org.scalatest.prop.TableDrivenPropertyChecks
+import org.scalatest.{FreeSpec, Matchers}
+
+final class LocalUniverseSpec extends FreeSpec {
 
   import LocalUniverseSpec._
 
-  "import endpoint" should "accept a Zip file" in { service =>
-    forAll (ValidFilenamePrefixes) { filenamePrefix =>
-      val packagePart = FileElement(
-        name = "file",
-        content = HelloWorldZip,
-        contentType = Some("application/zip"),
-        filename = Some(s"$filenamePrefix-digest.zip")
-      )
+  "import endpoint should" ignore {
 
-      val request = requestBuilder(ImportEndpoint)
-        .add(packagePart)
-        .buildFormPost(multipart = true)
+    "accept a Zip file" in {
+      forAll (ValidFilenamePrefixes) { filenamePrefix =>
+        val packagePart = FileElement(
+          name = "file",
+          content = HelloWorldZip,
+          contentType = Some("application/zip"),
+          filename = Some(s"$filenamePrefix-digest.zip")
+        )
 
-      val response = service(request)
-      assertResult(Status.Ok)(response.status)
-      assertSuccessJson("Import successful!")(response.contentString)
+        val request = CosmosClient.requestBuilder(ImportEndpoint)
+          .add(packagePart)
+          .buildFormPost(multipart = true)
+
+        val response = CosmosClient(request)
+        assertResult(Status.Ok)(response.status)
+        assertSuccessJson("Import successful!")(response.contentString)
+      }
     }
-  }
 
-  it should "not allow GET requests" in { service =>
-    val request = Request(Method.Get, s"/$ImportEndpoint")
-    val response = service(request)
-    assertResult(Status.NotFound)(response.status)
-  }
+    "not allow GET requests" in {
+      val request = Request(Method.Get, s"/$ImportEndpoint")
+      val response = CosmosClient(request)
+      assertResult(Status.NotFound)(response.status)
+    }
 
-  it should "only accept multipart requests" in { service =>
-    val request = requestBuilder(ImportEndpoint)
-      .buildPost(HelloWorldZip)
+    "only accept multipart requests" in {
+      val request = CosmosClient.requestBuilder(ImportEndpoint)
+        .buildPost(HelloWorldZip)
 
-    val response = service(request)
-    assertResult(Status.BadRequest)(response.status)
-  }
-
-  it should "require a form-data field named 'file' be present" in { service =>
-    val request = requestBuilder(ImportEndpoint)
-      .add(FileElement(name = "import", content = HelloWorldZip))
-      .buildFormPost(multipart = true)
-
-    val response = service(request)
-    assertResult(Status.BadRequest)(response.status)
-  }
-
-  it should "require the 'file' field to have an application/zip Content-type" in { service =>
-    val request = requestBuilder(ImportEndpoint)
-      .add(FileElement(name = "file", content = HelloWorldZip))
-      .buildFormPost(multipart = true)
-
-    val response = service(request)
-    assertResult(Status.BadRequest)(response.status)
-  }
-
-  it should "require an uploaded filename matching <package>-<version>-<digest>.zip" in { service =>
-    forAll (InvalidFilenames) { filename =>
-      val packagePart = FileElement(
-        name = "file",
-        content = HelloWorldZip,
-        contentType = Some("application/zip"),
-        filename = filename
-      )
-      val request = requestBuilder(ImportEndpoint)
-        .add(packagePart)
-        .buildFormPost(multipart = true)
-
-      val response = service(request)
+      val response = CosmosClient(request)
       assertResult(Status.BadRequest)(response.status)
     }
-  }
 
-  it should "require the package to be a non-empty Zip archive" in { service =>
-    val fileContents = Table("content", Buf.Empty, HelloWorldPackageJson)
-    forAll (fileContents) { content =>
-      val packagePart = FileElement(
-        name = "file",
-        content = content,
-        contentType = Some("application/zip"),
-        filename = Some("package-1.2.3-digest.zip")
-      )
-      val request = requestBuilder(ImportEndpoint)
-        .add(packagePart)
+    "require a form-data field named 'file' be present" in {
+      val request = CosmosClient.requestBuilder(ImportEndpoint)
+        .add(FileElement(name = "import", content = HelloWorldZip))
         .buildFormPost(multipart = true)
 
-      val response = service(request)
+      val response = CosmosClient(request)
       assertResult(Status.BadRequest)(response.status)
     }
+
+    "require the 'file' field to have an application/zip Content-type" in {
+      val request = CosmosClient.requestBuilder(ImportEndpoint)
+        .add(FileElement(name = "file", content = HelloWorldZip))
+        .buildFormPost(multipart = true)
+
+      val response = CosmosClient(request)
+      assertResult(Status.BadRequest)(response.status)
+    }
+
+    "require an uploaded filename matching <package>-<version>-<digest>.zip" in {
+      forAll (InvalidFilenames) { filename =>
+        val packagePart = FileElement(
+          name = "file",
+          content = HelloWorldZip,
+          contentType = Some("application/zip"),
+          filename = filename
+        )
+        val request = CosmosClient.requestBuilder(ImportEndpoint)
+          .add(packagePart)
+          .buildFormPost(multipart = true)
+
+        val response = CosmosClient(request)
+        assertResult(Status.BadRequest)(response.status)
+      }
+    }
+
+    "require the package to be a non-empty Zip archive" in {
+      val fileContents = Table("content", Buf.Empty, HelloWorldPackageJson)
+      forAll (fileContents) { content =>
+        val packagePart = FileElement(
+          name = "file",
+          content = content,
+          contentType = Some("application/zip"),
+          filename = Some("package-1.2.3-digest.zip")
+        )
+        val request = CosmosClient.requestBuilder(ImportEndpoint)
+          .add(packagePart)
+          .buildFormPost(multipart = true)
+
+        val response = CosmosClient(request)
+        assertResult(Status.BadRequest)(response.status)
+      }
+    }
+
   }
 
 }
 
-object LocalUniverseSpec extends IntegrationSpec {
+object LocalUniverseSpec extends Matchers with TableDrivenPropertyChecks  {
 
   val ImportEndpoint = "package/import"
