+++ src/it/scala/com/mesosphere/cosmos/handler/CapabilitiesHandlerSpec.scala
@@ -1,51 +1,29 @@
 package com.mesosphere.cosmos.handler
 
-import java.nio.file.Files
-
 import cats.data.Xor
 import com.mesosphere.cosmos.circe.Decoders._
 import com.mesosphere.cosmos.http.MediaTypes
-import com.mesosphere.cosmos.model.{Capability, CapabilitiesResponse}
-import com.mesosphere.cosmos.{Cosmos, IntegrationSpec}
-import com.twitter.finagle.Service
-import com.twitter.finagle.http.{Request, Response, Status}
+import com.mesosphere.cosmos.model.{CapabilitiesResponse, Capability}
+import com.mesosphere.cosmos.test.CosmosIntegrationTestClient._
+import com.twitter.finagle.http.Status
 import io.circe.parse._
+import org.scalatest.FreeSpec
 
-final class CapabilitiesHandlerSpec extends IntegrationSpec {
+final class CapabilitiesHandlerSpec extends FreeSpec {
+  lazy val logger = org.slf4j.LoggerFactory.getLogger(classOf[CapabilitiesHandlerSpec])
 
-  val tmpDir = {
-    val tempDir = Files.createTempDirectory("cosmos-CapabilitiesHandlerSpec")
-    val file = tempDir.toFile
-    Runtime.getRuntime.addShutdownHook(new Thread() {
-      override def run(): Unit = {
-        if (!com.twitter.io.Files.delete(file)) {
-          logger.warn("unable to cleanup temp dir: {}", file.getAbsolutePath)
-        }
-      }
-    })
-    val value = file.getAbsolutePath
-    logger.info("Setting com.mesosphere.cosmos.dataDir={}", value)
-    System.setProperty("com.mesosphere.cosmos.dataDir", value)
-    tempDir
+  "The capabilites handler should" - {
+    "be return a document" in {
+      val request = CosmosClient.requestBuilder("capabilities")
+        .setHeader("Accept", MediaTypes.CapabilitiesResponse.show)
+        .buildGet()
+      val response = CosmosClient(request)
+      val responseBody = response.contentString
+      logger.info("responseBody = {}", responseBody)
+      assertResult(Status.Ok)(response.status)
+      assertResult(MediaTypes.CapabilitiesResponse.show)(response.headerMap("Content-Type"))
+      val Xor.Right(body) = decode[CapabilitiesResponse](responseBody)
+      assertResult(CapabilitiesResponse(List(Capability("PACKAGE_MANAGEMENT"))))(body)
+    }
   }
-
-  val service = Cosmos.service
-
-  override def createService: Service[Request, Response] = {
-    service
-  }
-
-  "The capabilites handler" should "be return a document" in { service =>
-    val request = requestBuilder("capabilities")
-      .setHeader("Accept", MediaTypes.CapabilitiesResponse.show)
-      .buildGet()
-    val response = service(request)
-    val responseBody = response.contentString
-    logger.info("responseBody = {}", responseBody)
-    assertResult(Status.Ok)(response.status)
-    assertResult(MediaTypes.CapabilitiesResponse.show)(response.headerMap("Content-Type"))
-    val Xor.Right(body) = decode[CapabilitiesResponse](responseBody)
-    assertResult(CapabilitiesResponse(List(Capability("PACKAGE_MANAGEMENT"))))(body)
-  }
-
 }
